+++ Data/DList.hs
@@ -1,4 +1,5 @@
 {-# OPTIONS_GHC -Wall -O2 #-}
+{-# OPTIONS_HADDOCK prune #-}
 {-# LANGUAGE CPP #-}
 
 -----------------------------------------------------------------------------
@@ -17,7 +18,7 @@
 
 module Data.DList (
 
-   DList
+   DList(..)
 
   -- * Construction
   ,fromList      -- :: [a] -> DList a
@@ -38,6 +39,8 @@ module Data.DList (
   ,unfoldr       -- :: (b -> Maybe (a, b)) -> b -> DList a
   ,foldr         -- :: (a -> b -> b) -> b -> DList a -> b
   ,map           -- :: (a -> b) -> DList a -> DList b
+
+  ,maybeReturn
 
   ) where
 
@@ -80,6 +83,8 @@ import qualified Control.Applicative (empty)
 -- >       flatten (Branch x y) = flatten x >> flatten y
 --
 newtype DList a = DL { unDL :: [a] -> [a] }
+{-# DEPRECATED DL "It will be removed in dlist-v0.7 (see <https://github.com/spl/dlist/issues/4 #4>)." #-}
+{-# DEPRECATED unDL "It will be removed in dlist-v0.7. Use 'apply' instead." #-}
 
 -- | Converting a normal list to a dlist
 fromList    :: [a] -> DList a
@@ -258,4 +263,8 @@ instance Foldable DList where
   foldr' f x  = F.foldr' f x . toList
   {-# INLINE foldr' #-}
 #endif
+
+maybeReturn :: MonadPlus m => Maybe a -> m a
+maybeReturn = maybe mzero return
+{-# DEPRECATED maybeReturn "It will be removed in dlist-v0.7." #-}