+++ ChangeLog.md
@@ -11,6 +11,7 @@ Version 0.6 (*expected*)
 * Add instances: `Eq`, `Ord`, `Read`, `Show`, `Alternative`, and `Foldable`
 * Make DList abstract (see [#4](https://github.com/spl/dlist/issues/4))
 * Add `apply` to use instead of `unDL`
+* Remove `maybeReturn` which is not directly relevant to dlists
 * Stop supporting `base < 2`
 * Update tests to run `cabal test` using parallel QuickCheck (`pqc`)
 * Add scripts for running `hpc`