+++ ChangeLog.md
@@ -9,9 +9,9 @@ Version 0.6 (*expected*)
 * Migrate repository from http://code.haskell.org/~dons/code/dlist/ to
   https://github.com/spl/dlist
 * Add instances: `Eq`, `Ord`, `Read`, `Show`, `Alternative`, and `Foldable`
-* Make DList abstract (see [#4](https://github.com/spl/dlist/issues/4))
+* Deprecate DList constructor to make it abstract (see [#4](https://github.com/spl/dlist/issues/4))
 * Add `apply` to use instead of `unDL`
-* Remove `maybeReturn` which is not directly relevant to dlists
+* Deprecate `maybeReturn` which is not directly relevant to dlists
 * Stop supporting `base < 2`
 * Update tests to run `cabal test` using parallel QuickCheck (`pqc`)
 * Add scripts for running `hpc`